% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary_dat}
\alias{boundary_dat}
\title{Boundary data for Southern Asia}
\format{
An \code{sf} data frame with polygon geometries.
}
\source{
GADM database \url{https://gadm.org/}
}
\usage{
boundary_dat
}
\description{
An \code{sf} object containing boundary information for Southern Asia,
used for extracting country boundaries (e.g., Nepal).
}
\keyword{datasets}
