% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa_posthoc.R
\name{calc_feat_rep}
\alias{calc_feat_rep}
\title{Calculate feature representation}
\usage{
calc_feat_rep(prob.all, elsa_result, elsa_raster, feat_df, ELSA_text)
}
\arguments{
\item{prob.all}{A \code{prioritizr} problem. Needed as an input for the \code{prioritizr} function \code{eval_feature_representation_summary()}}

\item{elsa_result}{A \code{prioritizr} solution. Needed as an input for the \code{prioritizr} function \code{eval_feature_representation_summary()}}

\item{elsa_raster}{A \code{SpatRaster} containing label information of the the zones used to obtain solution. Has no zero values.}

\item{feat_df}{A \code{df} containing information on the features, including columns "label" and "theme"}

\item{ELSA_text}{A \code{tbl_df} containing the translations for the displayed text in the plot. Needs}
}
\value{
A \code{list} with two \code{tbl_df}, one with the raw data and the second with the data prepared for plotting.
}
\description{
This function calculates the feature representation of a single feature per zone comapred to all possible PUs for this feature across all zones.
E.g. if 60 out of 100 possible PUs are selected for zone 1 and 50 out of 100 possible PUs are selected for zone 2:
This DOES not mean that the overall feature representation in the planning region is 60\% (60/100) + 50\% (50/100) (= 110\%, so > 100\%)
Instead we need to get the representation in the zone versus all possible planning units in the whole region.
This means that if in our case zones do not share any overlapping possible area prior to prioritisation, we would have 200 possible planning units overall for the feature in the planning region.
The representation across the whole planning region would then be: 60/200 + 50/200, so 110/200 (= 55\%)
}
