% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{check_existing_ee_export_task}
\alias{check_existing_ee_export_task}
\title{Check for existing Earth Engine export task}
\usage{
check_existing_ee_export_task(ee, description)
}
\arguments{
\item{ee}{The Earth Engine Python module}

\item{description}{Description of the export task}
}
\value{
TRUE if a task with matching description is found running or ready
}
\description{
Check for existing Earth Engine export task
}
\keyword{internal}
