% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_points_polygon}
\alias{convert_points_polygon}
\title{Convert Points to Buffered Polygons Based on Area}
\usage{
convert_points_polygon(
  wdpa_layer,
  area_attr = "REP_AREA",
  area_crs = "ESRI:54009",
  nQuadSegs = 50,
  append_sf = TRUE,
  area_multiplier = 1e+06
)
}
\arguments{
\item{wdpa_layer}{An \code{sf} object with point or multipoint geometries and an area field.}

\item{area_attr}{Name of the area attribute column (default = "REP_AREA").}

\item{area_crs}{CRS to use for buffering (default = 'ESRI:54009' = World Mollweide).}

\item{nQuadSegs}{Number of segments per circle quadrant used in buffering (default = 50).}

\item{append_sf}{Logical. If TRUE, append buffered geometries to existing polygons (default = TRUE).}

\item{area_multiplier}{Numeric. Scaling multiplier for area units (e.g., 1e6 for km^2).}
}
\value{
An \code{sf} object with polygons for both original and buffered geometries.
}
\description{
Converts point geometries to circular polygon buffers based on an area attribute.
}
