% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_global_raster}
\alias{crop_global_raster}
\title{Crop a Global Raster to the Extent of Planning Units}
\usage{
crop_global_raster(raster_in, pus, threads = TRUE)
}
\arguments{
\item{raster_in}{SpatRaster. A large input raster (e.g., global dataset).}

\item{pus}{SpatRaster. Planning units raster used to define the target extent.}

\item{threads}{Optional method to use multi-core processing - to speed on some \code{terra} functions (default: \code{TRUE}).}
}
\value{
A cropped SpatRaster with the same CRS as \code{raster_in} and extent matching the reprojected \code{pus}.
}
\description{
This function crops a large global raster to the spatial extent of a planning units raster (\code{pus}).
It reprojects the \code{pus} extent to the coordinate reference system of the input global raster
to ensure accurate cropping. This is useful for pre-processing global inputs before using
functions like \code{make_normalised_raster()} to reduce processing time.
}
\examples{
\dontrun{
cropped <- crop_global_raster(global_raster, pus) |>
  elsar::make_normalised_raster(pus = pus, iso3 = "KEN")
}
}
