% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{current_pas_sf}
\alias{current_pas_sf}
\title{Current Protected Areas}
\format{
An \code{sf} data frame with polygon geometries.
}
\source{
UNEP-WCMC and IUCN WDPA \url{https://www.protectedplanet.net/}
}
\usage{
current_pas_sf
}
\description{
An \code{sf} object containing current protected areas from the
World Database on Protected Areas (WDPA).
}
\keyword{datasets}
