% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{download_from_drive}
\alias{download_from_drive}
\title{Download files from Google Drive}
\usage{
download_from_drive(folder, prefix, dest)
}
\arguments{
\item{folder}{Drive folder name}

\item{prefix}{File prefix to search for}

\item{dest}{Local destination folder}
}
\value{
Vector of downloaded file paths
}
\description{
Download files from Google Drive
}
\keyword{internal}
