% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa_plotting.R
\name{elsa_plot_output_map}
\alias{elsa_plot_output_map}
\title{Create a Map of the ELSA output to download with GIS layers}
\usage{
elsa_plot_output_map(elsa_raster, ELSA_text, pu0)
}
\arguments{
\item{elsa_raster}{A \code{SpatRaster}with one layer that contains different categories for the different zones.}

\item{ELSA_text}{A \code{tbl_df} containing the translations for the displayed text in the plot.}

\item{pu0}{A \code{SpatRaster} of all possible planning units in the planning region (one layer)}
}
\value{
A ggplot object with a map of the prioritization output
}
\description{
This function gives a quick way of getting a map immediately when downloading the output GIS layers.
}
