% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa_plotting.R
\name{elsa_plot_representation}
\alias{elsa_plot_representation}
\title{Create a Bar Plot of Feature Representation}
\usage{
elsa_plot_representation(
  feature_rep_tabl,
  input,
  ELSA_text,
  pal.elsa = tibble::tibble(colour = c("#4daf4a", "#984ea3", "#377eb8", "#adaf4a",
    "#FFFFFF00"), category = c("Protect", "Restore", "Manage", "PR", "Do_Nothing"))
)
}
\arguments{
\item{feature_rep_tabl}{A dataframe containing representation values for each feature.
Columns should include numeric data for each feature and action.}

\item{input}{The input object of an R Shiny app containing budget information for each action.
This is used to label the budget percentage of each action in the plot legend.}

\item{ELSA_text}{A \code{tbl_df} containing the translations for the displayed text in the plot.}

\item{pal.elsa}{A \code{tibble} with palette information for plot}
}
\value{
A ggplot object with a stacked bar chart showing feature representation percentages.
}
\description{
This function generates a stacked bar plot displaying the representation of each input
feature, broken up into specified actions (e.g., "Protect," "Restore," "Manage,"
and optionally "Urban Greening").
}
