% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsafunctions-package.R
\docType{package}
\name{elsaFunctions-package}
\alias{elsaFunctions}
\alias{elsaFunctions-package}
\title{elsaFunctions: Functions and Tools For Creating ELSA Tier 1 Data Stacks}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Scott Atkinson \email{scott.atkinson@undp.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
