% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar-package.R
\docType{package}
\name{elsar-package}
\alias{elsar}
\alias{elsar-package}
\title{elsar: Functions and Tools For Creating ELSA Tier 1 Data Stacks}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elsa-undp.github.io/elsar/}
}

}
\author{
\strong{Maintainer}: Scott Atkinson \email{scott.atkinson@undp.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Sandra Neubert \email{s.neubert@uq.edu.au} (\href{https://orcid.org/0000-0002-3112-4116}{ORCID})
}

}
\keyword{internal}
