% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar_plot_feature.R
\name{elsar_plot_feature}
\alias{elsar_plot_feature}
\title{Create a map of individual features}
\usage{
elsar_plot_feature(
  raster_in,
  pus,
  legend_title = NULL,
  color_map = "viridis",
  invert_palette = FALSE,
  figure_path = NULL,
  no_legend = FALSE
)
}
\arguments{
\item{raster_in}{The \code{SpatRaster} file to be plotted.}

\item{pus}{\code{SpatVector} The planning units (PUs) to use as outline}

\item{legend_title}{An optional legend title.}

\item{color_map}{The name of the \code{viridis} palette to be used. Default is "viridis".}

\item{figure_path}{An optional path to save the produced figures}

\item{no_legend}{Logical. TRUE means no legend will be shown}
}
\value{
A ggplot object with a map of the feature
}
\description{
Create a map of individual features
}
\examples{
boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd"
  )

pus <- make_planning_units(
  boundary_proj = boundary_proj,
  pu_size = NULL,
  pu_threshold = 8.5e5,
  limit_to_mainland = FALSE
  )

wad_dat <- get_wad_data()

wadOut <- make_normalised_raster(
  raster_in = wad_dat,
  pus = pus,
  iso3 = "NPL"
  )

elsar_plot_feature(
raster_in = wadOut,
pus = pus,
legend_title = "wad"
)

}
