% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar_plot_static_vector.R
\name{elsar_plot_static_vector}
\alias{elsar_plot_static_vector}
\title{Plot vector}
\usage{
elsar_plot_static_vector(
  sf_in,
  data_type = "d",
  col_interest = NULL,
  custom_palette = NULL,
  color_map = "viridis",
  background = NULL,
  extend_background = 0.05,
  plot_title = "",
  legend_title = "",
  legend_labels = NULL,
  expand_plot = FALSE
)
}
\arguments{
\item{sf_in}{An\code{sf} object containing the data to be plotted.}

\item{data_type}{Whether the data is discrete ("d") or continuous ("c")}

\item{col_interest}{A list of column names to include in the plot. If specified, only these columns will be used to colour the plot.}

\item{custom_palette}{An optional custom palette for plotting. Default uses the \code{viridis} package.}

\item{color_map}{The name of the \code{viridis} palette to be used. Default is "viridis".}

\item{background}{Requires a \code{SpatRaster} input (preferably with the same data as \code{raster_in}) to plot the background data.}

\item{extend_background}{A numerical value that allows to extent the background beyond the extent of \code{raster_in}. If extend_background <= 1, the lat and lon extend will be extended by the ratio provided (e.g. 0.05 will extend it by 5\%). If extend_background > 1 all sides will be extended by the absolute value provided.}

\item{plot_title}{An optional plot title.}

\item{legend_title}{An optional legend title.}

\item{legend_labels}{A vector of strings containing the labels to use for legend values.}

\item{expand_plot}{logical. Whether to expand the plot, so there is not space betwen border and plot.}
}
\value{
A ggplot object.
}
\description{
Plot vector
}
