% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{ensure_drive_folder}
\alias{ensure_drive_folder}
\title{Ensure Google Drive folder exists}
\usage{
ensure_drive_folder(folder_name)
}
\arguments{
\item{folder_name}{Name of the Google Drive folder to check or create}
}
\description{
Ensure Google Drive folder exists
}
\keyword{internal}
