% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_filename_filetype}
\alias{extract_filename_filetype}
\title{Extract Filename and Filetype from Directory}
\usage{
extract_filename_filetype(data_name, file_path)
}
\arguments{
\item{data_name}{Character. Name pattern of the file to search.}

\item{file_path}{Character. Path to the folder.}
}
\value{
A named list with elements: \code{filename} and \code{filetype}.
}
\description{
Extracts the base filename and file extension for a given \code{data_name} in a specified folder.
}
