% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_features.R
\name{get_all_features}
\alias{get_all_features}
\title{Wrapper function to get a raster file with all wanted features}
\usage{
get_all_features(feature_list, path_in, pus, iso3, output_path = NULL, ...)
}
\arguments{
\item{feature_list}{A \code{list} with all features that are wanted for the downstream ELSA workflow.}

\item{path_in}{A common path where all the data can be found.}

\item{pus}{A \code{SpatRaster} file that contains the reference spatial extent, crs etc.in form of the planning units}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{output_path}{An optional output path for all created files.}

\item{...}{Additional attributes needed for some functions if wanted.}
}
\value{
A \code{SpatRaster} with the length of the feature_list + 1 since the planning units are the first layer.
}
\description{
Wrapper function to get a raster file with all wanted features
}
\examples{
\dontrun{
feature_list <- c("fml", "cropsuit", "fi")

all_feats <- get_all_features(feature_list = feature_list,
                              path_in = "<yourpath>",
                              pus = pus,
                              iso3 = "NPL")
}
}
