% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coverage}
\alias{get_coverage}
\title{Calculate Areal Coverage of a Zone Relative to Planning Units}
\usage{
get_coverage(zone_layer, pu_layer)
}
\arguments{
\item{zone_layer}{A binary \code{SpatRaster} representing a target zone.}

\item{pu_layer}{A binary \code{SpatRaster} representing planning units.}
}
\value{
Numeric. Percentage of zone coverage across PUs.
}
\description{
Calculates percentage of a zone layer relative to the total area of planning units.
}
