% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa.R
\name{get_elsa_coverage}
\alias{get_elsa_coverage}
\title{Calculate the Percentage Area Coverage for a Given Zone}
\usage{
get_elsa_coverage(zone_layer, pu_layer)
}
\arguments{
\item{zone_layer}{A \code{terra} raster object representing the specific zone (e.g., protected or managed areas) for which
the coverage percentage will be calculated.}

\item{pu_layer}{A \code{terra} raster object representing the total planning units (PU) layer that defines the entire area of interest.}
}
\value{
A numeric value representing the percentage coverage of the \code{zone_layer} relative to the \code{pu_layer}.
}
\description{
This function calculates the percentage of the total area covered by a specific zone (e.g., protected areas, managed areas)
relative to the entire planning unit (PU) layer. It returns the areal coverage as a percentage.
}
\details{
The function uses \code{terra::global()} to calculate the sum of values in both the zone and PU layers,
ignoring missing values (\code{NA}s), and computes the percentage of the zone's total area relative to the PU's total area.
}
\examples{
\dontrun{
# Calculate coverage for protected areas
get_coverage(PA, pu)
# Calculate coverage for protected zones
get_coverage(zone_protect, pu)
# Calculate coverage for managed zones
get_coverage(zone_manage, pu)
}
}
\seealso{
\code{\link[terra:global]{terra::global()}} which is used to sum the values in the raster layers.
}
