% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa.R
\name{get_elsa_min_lockin_target}
\alias{get_elsa_min_lockin_target}
\title{Get Minimum Lock-in Targets for Zones}
\usage{
get_elsa_min_lockin_target(lockin, input, pu)
}
\arguments{
\item{lockin}{A raster or list of rasters representing the locked-in areas for each zone.}

\item{input}{A list or dataframe containing the target values for each zone (e.g., \code{zone_1_target}, \code{zone_2_target}, etc.).}

\item{pu}{A raster layer representing the planning unit used for calculating coverage.}
}
\value{
A \code{tibble} containing the adjusted target values for each zone.
}
\description{
This function calculates the minimum lock-in targets for different zones (e.g., protect, restore, manage, and optionally green).
It compares the input targets with the minimum coverage needed for each zone and ensures that the target is not set below this threshold.
}
