% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa.R
\name{get_elsa_target}
\alias{get_elsa_target}
\title{Get country-specific target value based on the total area of planning units (PU)}
\usage{
get_elsa_target(PU = NULL, target = NULL)
}
\arguments{
\item{PU}{A \code{terra} raster or object representing the planning units, where the total area is calculated.}

\item{target}{A numeric value representing the percentage target (e.g., 10 for 10\%).}
}
\value{
A numeric value representing the target area. If the calculated target is zero,
the function returns a very small value (1e-4) to ensure a non-zero target.
}
\description{
This function calculates a target value by taking a percentage of the total area
represented by the planning units (PU). The result is rounded to the nearest whole number,
and if the calculated target is zero, a small value (1e-4) is returned to avoid issues with zero targets.
}
\examples{
\dontrun{
# Example usage with a raster of planning units and a target of 10\%
get_target(PU = my_planning_units_raster, target = 10)
}
}
