% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn_ecosystems.R
\name{get_iucn_ecosystems}
\alias{get_iucn_ecosystems}
\title{Extract IUCN GET Ecosystems Vector Layers}
\usage{
get_iucn_ecosystems(
  iucn_get_directory,
  iso3,
  pus,
  include_minor_occurence = TRUE,
  iucn_get_prefixes = NULL,
  output_path = NULL
)
}
\arguments{
\item{iucn_get_directory}{Character. Path to the directory containing \code{.gpkg} files.}

\item{iso3}{Character. ISO3 country code (e.g., "KEN"), used in optional output naming.}

\item{pus}{SpatRaster. Planning units raster used to define spatial extent and CRS.}

\item{include_minor_occurence}{Logical. If FALSE, features marked as "minor" occurrence will be excluded (default = TRUE).}

\item{output_path}{Optional character. If provided, writes the merged ecosystem layer as a GeoPackage.}

\item{prefixes}{Character vector of filename prefixes to include (e.g., \code{c("T", "TF", "FM")}).
Must match one or more of: \code{"F"}, \code{"FM"}, \code{"M"}, \code{"MFT"}, \code{"MT"}, \code{"S"}, \code{"SF"}, \code{"SM"}, \code{"T"}, \code{"TF"}.
If \code{NULL} (default), all \code{.gpkg} files in the directory will be included.}
}
\value{
An \code{sf} object containing all filtered IUCN GET ecosystem polygons intersecting the PUs extent.
}
\description{
This function reads and merges \code{.gpkg} vector files representing IUCN GET ecosystems
from a specified directory. Optionally, it filters files by a set of filename prefixes
(e.g., "T", "TF", "FM", etc.), restricts features to those intersecting the extent
of a planning units raster, and filters out minor occurrence ecosystems. If \code{output_path}
is specified, the result is written to a GeoPackage.
}
\examples{
\dontrun{
pus <- terra::rast("data/pus_raster.tif")
ecosystems <- get_iucn_ecosystems(
  iucn_get_directory = "data/iucn_layers",
  iso3 = "KEN",
  pus = pus,
  prefixes = c("T", "TF"),
  include_minor_occurence = FALSE,
  output_path = "outputs"
)
}
}
