% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa.R
\name{get_min_lockin_target}
\alias{get_min_lockin_target}
\title{Get Minimum Lock-in Targets for Zones}
\usage{
get_min_lockin_target(lockin, input, pu)
}
\arguments{
\item{lockin}{A raster or list of rasters representing locked-in areas for
each zone.}

\item{input}{A list or dataframe containing target values for each zone
(e.g., \code{zone_1_target}, \code{zone_2_target}, etc.).}

\item{pu}{A raster layer representing the planning unit used for calculating
coverage.}
}
\value{
A \code{tibble} containing the adjusted target values for each zone.
}
\description{
Calculates the minimum lock-in targets for different zones (e.g., protect,
restore, manage, and optionally green). Compares input targets with the
minimum coverage needed for each zone and ensures targets are not set below
this threshold.
}
