% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_underrepresented_ecosystems}
\alias{get_underrepresented_ecosystems}
\title{Calculate Underrepresented Ecosystems Across a Set}
\usage{
get_underrepresented_ecosystems(
  x,
  pus,
  iso3 = iso3,
  invert = FALSE,
  rescaled = TRUE,
  fun = mean
)
}
\arguments{
\item{x}{An \code{sf} or \code{data.frame} with one row per ecosystem polygon and a \code{target} column.}

\item{pus}{A \code{SpatRaster} representing the planning units grid.}

\item{iso3}{ISO3 country code used for naming.}

\item{invert}{Logical. If TRUE, invert values before normalization (default = FALSE).}

\item{rescaled}{Logical. If TRUE, normalize result to \link{0, 1} (default = TRUE).}

\item{fun}{Function. Aggregation function to apply across rasters (default = mean).}
}
\value{
A normalized \code{SpatRaster} showing weighted average underrepresentation.
}
\description{
Applies a weighted averaging function across a set of ecosystem coverage rasters.
}
