% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_wad_data}
\alias{get_wad_data}
\title{Get WAD (World Atlas of Desertification) subset data}
\usage{
get_wad_data()
}
\value{
A \code{SpatRaster} object containing the WAD subset data.
}
\description{
Loads a subset of WAD data saved as a TIF file from the package's
extdata directory.
}
\examples{
\dontrun{
wad <- get_wad_data()
}
}
