% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not in operator}
\usage{
x \%ni\% df
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{df}{vector, data frame column, or similar: the values to be matched against.}
}
\value{
A logical vector, TRUE where \code{x} is not in \code{df}.
}
\description{
A negation of \code{\%in\%}, for checking if elements are \emph{not} in another object.
}
\examples{
1 \%ni\% 1:5
6 \%ni\% 1:5

}
