% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load different types of data}
\usage{
load_data(
  file_name,
  file_path = NULL,
  file_lyr = NULL,
  file_type,
  db_info,
  iso3_column = "iso3cd",
  iso3
)
}
\arguments{
\item{file_name}{character of the file name. Needs to contain the file type ending (e.g. \code{.tif}) if loading from a local source.}

\item{file_path}{path where local file is stored. Needs to be \code{NULL} when using postgres.}

\item{file_lyr}{optional. Layer information of data.}

\item{file_type}{character of file type. Current options are: "postgres", "shp", "gpkg", "geojson", "tif", "tiff", "grd", "gri", "nc", "hdf"}

\item{db_info}{list in the style of a dictionary. Only needed when file_type = "postgres". Needs to have the following structure and information: postgres_dict <- c(host = "\if{html}{\out{<yourhost>}}", dbname ="\if{html}{\out{<yourdbname>}}", port = \if{html}{\out{<portNumber>}}, user = "\if{html}{\out{<yourusername>}}", password = "\if{html}{\out{<yourpassword>}}")}

\item{iso3_column}{Only relevant when \code{file_type} "postgres" is selected. A string of the name of where iso3 information can be found in a dataset.}

\item{iso3}{The iso3 country code (character) of the country of interest.}
}
\value{
The loaded data either as a \code{SpatRaster} or \code{sf} object
}
\description{
Load different types of data
}
\examples{
\dontrun{
load_tif <- load_data(file_name = "pu_nepal_450m.tif", file_path = localPath
                          file_type = "tif")

load_geojson <- load_data(file_name = "nepal.geojson", file_path = localPath,
                          file_type = "geojson")

postgres_dict <- c(host = "<yourhost>",
                   dbname ="<yourdbname>",
                   port = <portNumber>,
                   user = "<yourusername>",
                   password = "<yourpassword>")
load_postgres <- load_data(file_name = "bnda_simplified",file_type = "postgres",
                           db_info = postgres_dict,
                           iso3_column = "iso3cd",
                           iso3 = "NPL")
}
}
