% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-elsa_plotting.R
\name{make_categorical_raster}
\alias{make_categorical_raster}
\title{Function to create a categorical raster based on an input raster}
\usage{
make_categorical_raster(in_rast, elsa_categories, pal.elsa, ELSA_text)
}
\arguments{
\item{in_rast}{A \code{SpatRaster} that is to be divided into categories.}

\item{elsa_categories}{A \code{df} containing columns \code{action} and \code{category}}

\item{pal.elsa}{A \code{df} with palette values for ELSA categories. Needs to contain column \code{category}}

\item{ELSA_text}{A \code{tbl_df} containing the translations for the displayed text in the plot.}
}
\value{
A \code{SpatRaster} with categorical data that includes category values, colors, and labels.
}
\description{
This function converts an input \code{SpatRaster} into a categorical raster. The input raster
is processed to remove zero values, and categorical data are derived based on the number
of layers and layer names. Additionally, the categories are linked to a global color
palette and associated labels.
}
