% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_water_areas.R
\name{make_clean_water_areas}
\alias{make_clean_water_areas}
\title{This function processes a realised clean water raster by aligning it to a
country's planning unit raster and normalizing the values within the planning unit.}
\usage{
make_clean_water_areas(clean_water_input, pu, iso3, output_dir = NULL)
}
\arguments{
\item{clean_water_input}{Character. Path to the input clean water raster.}

\item{pu}{Character. Path to the planning unit raster.}

\item{iso3}{Character. ISO3 country code to identify the country in the output file name.}

\item{output_dir}{Character. Directory where the output raster will be saved. If NULL, the result is not saved.}
}
\value{
A \code{SpatRaster} object of the normalized clean water raster.
If \code{output_dir} is not NULL, the result is also saved as a Cloud Optimized GeoTIFF (COG).
}
\description{
This function processes a realised clean water raster by aligning it to a
country's planning unit raster and normalizing the values within the planning unit.
}
\examples{
make_clean_water_areas(
  clean_water_input = "path/to/clean_water.tif",
  pu = "path/to/pu_raster.tif",
  iso3 = "HTI",
  output_dir = "path/to/output/directory"
)
}
