% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_custom_projection.R
\name{make_custom_projection}
\alias{make_custom_projection}
\title{Create a custom projection based on the planning region}
\usage{
make_custom_projection(
  boundary,
  output_path = NULL,
  iso3_column = "iso3cd",
  iso3 = NULL
)
}
\arguments{
\item{boundary}{\code{sf} object of the boundary of the planning region. Should match iso3 country code.}

\item{output_path}{An optional output path for the created file.}

\item{iso3_column}{A string of the name of where iso3 information can be found in a dataset.}

\item{iso3}{The iso3 country code (character) of the country of interest.}
}
\value{
A \code{wkt} file centred on the planning region
}
\description{
Create a custom projection based on the planning region
}
\examples{
boundary <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd"
)

wkt <- make_custom_projection(boundary = boundary, iso3 = "NPL")
}
