% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest_integrity.R
\name{make_forest_integrity}
\alias{make_forest_integrity}
\title{Create the forest integrity index using flii and fsii data}
\usage{
make_forest_integrity(raster_flii, raster_fsii, pus, iso3, output_path = NULL)
}
\arguments{
\item{raster_flii}{A \code{SpatRaster} file with the flii information.}

\item{raster_fsii}{A \code{SpatRaster} file with the fsii information.}

\item{pus}{A \code{SpatRaster} file that contains the reference spatial extent, crs etc.in form of the planning units}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{output_path}{An optional output path for the created file.}
}
\value{
A \code{SpatRaster} file of the forest integrity index that has been aligned and normalised
}
\description{
Create the forest integrity index using flii and fsii data
}
\examples{
\dontrun{
forest_integrity <- make_forest_integrity(raster_flii = raster_flii, raster_fsii = raster_fsii, pus = pus)
}
}
