% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_planning_units.R
\name{make_planning_units}
\alias{make_planning_units}
\title{Function to create planning units and custom projection}
\usage{
make_planning_units(
  boundary_src,
  boundary_lyr = NULL,
  dataPath,
  input_type = "sf",
  pu_size = NULL,
  pu_threshold = 850000,
  limit_to_mainland = FALSE,
  iso3_column = "iso_sov1",
  iso3,
  outputPath = NULL
)
}
\arguments{
\item{boundary_src}{A string of the data name and file type}

\item{boundary_lyr}{The layer used within the data (character or numeric).}

\item{dataPath}{A string of the path where the data is saved. This is also where the custom projection is saved.}

\item{input_type}{A string that is either "sf", "postgres" or "raster" (default is "sf").}

\item{pu_size}{A way to define a custom planning unit size. Can be NULL to use default settings that generate planning units as small as possible whilst still being computationally efficient.}

\item{pu_threshold}{An integer value that gives a maximum number of PUs. The default (\code{8.5e5}) is set based on \code{prioritizr} processing time, network transfer time and solver time.}

\item{limit_to_mainland}{A logical that determines whether planning units should only be created for mainland area (\code{FALSE}; default) or not (\code{TRUE})}

\item{iso3_column}{Only relevant when \code{ìnput_type} "postgres" is selected. A string of the name of where iso3 information can be found in a dataset.}

\item{iso3}{The iso3 country code (character) of the country of interest.}

\item{outputPath}{An optional output path for the created file. If none is provided, \code{dataPath} will be used.}
}
\value{
A raster (\code{.tif}) file with the planning unit information for the chosen planning region.
}
\description{
This function creates planning units for a spatial prioritisation problem in a raster format. It also creates a custom projection (\code{wkt} file) centred on the planning region based on the Mollweide projection.
}
\examples{
\dontrun{
pus_nepal <- make_planning_units(
  boundary_src = "pu_nepal_450m.tif",
  boundary_lyr = NULL,
  dataPath = "localDataPath/nepal_reference",
  input_type = "raster",
  iso3 = "NPL"
)
}
}
