% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster_general.R
\name{make_raster_general}
\alias{make_raster_general}
\title{Function to extract general raster data}
\usage{
make_raster_general(
  rasterIn,
  rasterRef,
  rasterMask,
  iso3,
  nameOut,
  writeFile = TRUE
)
}
\arguments{
\item{rasterIn}{A raster file that contains the data to be put into right format}

\item{rasterRef}{A raster file that contains the reference spatial extent, crs etc.}

\item{rasterMask}{Calculate this from rasterRef?}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{nameOut}{A string with the data name that will be used for the output \code{tif}file}

\item{writeFile}{A logical command on whether the file will be saved rather than given as an output variable (Default: TRUE)}
}
\value{
A raster file that has been aligned and normalised
}
\description{
\code{make_raster_general()}allows you to align, normalise and save a raster file. This is applicable for certified forests, etc. (ADD LATER)
}
