% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mangroves_hti}
\alias{mangroves_hti}
\title{Data used for mangroves}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1139 rows and 2 columns.
}
\source{
https:// add link here
}
\usage{
mangroves_hti
}
\description{
\code{sf} file from WDPA database
}
\keyword{datasets}
