% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{merge_tiles_to_cog}
\alias{merge_tiles_to_cog}
\title{Merge downloaded tiles into a single COG}
\usage{
merge_tiles_to_cog(tiles, output_file, datatype = "FLT4S")
}
\arguments{
\item{tiles}{List of GeoTIFF file paths}

\item{output_file}{Final merged COG file path}

\item{datatype}{Output raster datatype}
}
\value{
A \code{SpatRaster} object
}
\description{
Merge downloaded tiles into a single COG
}
\keyword{internal}
