% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{start_gee_export_task}
\alias{start_gee_export_task}
\title{Start new Earth Engine export task}
\usage{
start_gee_export_task(
  ee,
  image,
  description,
  region,
  folder,
  file_name,
  scale = 10,
  max_pixels = 1e+13,
  file_format = "GeoTIFF"
)
}
\arguments{
\item{ee}{Earth Engine module}

\item{image}{EE Image object to export}

\item{description}{Description of the task}

\item{region}{Geometry to export}

\item{folder}{Google Drive folder name}

\item{file_name}{Filename prefix}

\item{scale}{Export resolution}

\item{max_pixels}{Max pixels allowed}

\item{file_format}{Export file format}
}
\description{
Start new Earth Engine export task
}
\keyword{internal}
