% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{wait_for_drive_export}
\alias{wait_for_drive_export}
\title{Wait for GEE Drive export to appear}
\usage{
wait_for_drive_export(folder, prefix, wait_time = 5)
}
\arguments{
\item{folder}{Drive folder name}

\item{prefix}{Filename prefix to search for}

\item{wait_time}{Max wait time in minutes}
}
\description{
Wait for GEE Drive export to appear
}
\keyword{internal}
