% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary_dat}
\alias{boundary_dat}
\title{Data used to extract the Nepal boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.
}
\source{
https:// add link here
}
\usage{
boundary_dat
}
\description{
\code{sf} file containing the Southern Asia boundary information
}
\keyword{datasets}
