% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{check_and_download_required_layers}
\alias{check_and_download_required_layers}
\title{Check and download GEE-sourced layers (e.g. LULC or Pasturelands)}
\usage{
check_and_download_required_layers(
  data_info,
  iso3,
  input_path,
  gee_project,
  boundary_proj,
  wait_time = 5
)
}
\arguments{
\item{data_info}{A dataframe containing \code{data_name} fields to match}

\item{iso3}{ISO3 country code}

\item{input_path}{Path where data files should exist or be downloaded to}

\item{gee_project}{Google Earth Engine project ID}

\item{boundary_proj}{\code{sf} object representing the country boundary}

\item{wait_time}{Wait time (in minutes) for Drive export}
}
\value{
NULL (used for side-effect of downloading)
}
\description{
This checks for required layer names from metadata, ensures they exist in the output path,
and prompts the user to download them from GEE if missing.
}
\keyword{internal}
