% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{cleanup_earthengine}
\alias{cleanup_earthengine}
\title{Clean up Earth Engine environment}
\usage{
cleanup_earthengine(env_info)
}
\arguments{
\item{env_info}{List returned from initialize_earthengine()}
}
\description{
Removes temporary conda environment if one was created
}
\keyword{internal}
