% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{current_pas_sf}
\alias{current_pas_sf}
\title{Data used for current protected areas}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 1 columns.
}
\source{
https:// add link here
}
\usage{
current_pas_sf
}
\description{
\code{current_pas_sf} file from wdpa database
}
\keyword{datasets}
