% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar-package.R
\docType{package}
\name{elsar-package}
\alias{elsar}
\alias{elsar-package}
\title{elsar: Functions and Tools For Creating ELSA Data Stacks}
\description{
Provides tools and data pipelines to support Essential Life Support Areas (ELSA) conservation planning. Includes functions to load and harmonize global spatial datasets, prepare raster layers for spatial prioritization, create zone layers (protect, restore, manage), and calculate coverage metrics over planning units.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elsa-undp.github.io/elsar/}
}

}
\author{
\strong{Maintainer}: Scott Atkinson \email{scott.atkinson@undp.org} (\href{https://orcid.org/0000-0001-5995-0199}{ORCID})

Authors:
\itemize{
  \item Sandra Neubert \email{s.neubert@uq.edu.au} (\href{https://orcid.org/0000-0002-3112-4116}{ORCID})
}

}
\keyword{internal}
