% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar_download_esri_lulc_data.R
\name{elsar_download_esri_lulc_data}
\alias{elsar_download_esri_lulc_data}
\title{Download and Process ESRI 10m LULC Data from Google Earth Engine}
\usage{
elsar_download_esri_lulc_data(
  boundary_layer,
  iso3,
  output_dir = here::here(),
  lulc_data_source = "projects/sat-io/open-datasets/landcover/ESRI_Global-LULC_10m_TS",
  gee_project = "unbl-misc",
  googledrive_folder = "gee_exports",
  wait_time = 5
)
}
\arguments{
\item{boundary_layer}{An \code{sf} object representing the national boundary.}

\item{iso3}{A character string representing the ISO3 country code (e.g., "USA").}

\item{output_dir}{A character string specifying the output directory for the final COG file.
Defaults to the project root directory using \code{here::here()}.}

\item{lulc_data_source}{A character string specifying the Earth Engine dataset ID for LULC data.
Defaults to the ESRI 10m LULC dataset.}

\item{gee_project}{A string specifying the GEE cloud project; default to \code{unbl-misc}. You must be a member of the project.}

\item{googledrive_folder}{A string specifying the folder in Google Drive that GEE exports will go to, or where files will be searched for. Default to \code{gee_exports}. This folder will be created if if does not exist.}

\item{wait_time}{A integer indicating the amount of time to wait for GEE exports to finish before the function stops gracefully.
Defaults to 5 (minutes).}
}
\value{
A Cloud-Optimized GeoTIFF file stored in the specified output directory.
}
\description{
This function downloads ESRI 10m Land Use Land Cover (LULC) data from Google Earth Engine (GEE)
for a specified national boundary. It exports the image to Google Drive, downloads the resulting
tiles, merges them into a single raster, and converts it into a Cloud-Optimized GeoTIFF (COG).
Temporary files and environments are cleaned up after processing.
}
\examples{
\dontrun{
my_boundary <- sf::st_read("my_boundary.gpkg")

lulc <- elsar_download_esri_lulc_data(boundary = my_boundary, iso3 = 'NPL')
}
}
