% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar_load_data.R
\name{elsar_load_data}
\alias{elsar_load_data}
\title{Load spatial or raster data from local files or Postgres}
\usage{
elsar_load_data(
  file_name = NULL,
  file_path = NULL,
  file_lyr = NULL,
  wkt_filter = NULL,
  db_info = NULL,
  pg_connection = NULL,
  drop3d = TRUE,
  iso3_column = NULL,
  iso3 = NULL
)
}
\arguments{
\item{file_name}{Character or NULL. File name with extension. Use NULL to load all files of a type in a folder.}

\item{file_path}{Character. Folder or file path.}

\item{file_lyr}{Character. Optional. Layer name for multi-layer files (e.g., GeoPackage or GDB).}

\item{wkt_filter}{\code{sf}, \code{SpatRaster}, or \code{SpatVector}, from which WKT geometry can be derived to spatially filter vector data when reading in.}

\item{db_info}{Named list. For Postgres: host, dbname, port, user, password.}

\item{pg_connection}{Named list. Alternative to db_info for Postgres.}

\item{drop3d}{Logical. Drop Z/M dimensions if TRUE. Default is TRUE.}

\item{iso3_column}{Character. Column name to filter by ISO3 code.}

\item{iso3}{Character. ISO3 code to filter vector data.}
}
\value{
An sf object or SpatRaster depending on the file type.
}
\description{
Automatically detects file type and loads spatial (\code{sf}) or raster (\code{SpatRaster}) data.
Supports local files (e.g., shapefiles, GeoPackages, TIFFs) and PostGIS tables.
}
\examples{
\dontrun{
# Load a single raster
r <- elsar_load_data(
  file_name = "lulc.tif",
  file_path = "data"
  )

# Load shapefile with ISO3 filter
shp <- elsar_load_data(
  file_name = "admin.shp",
  file_path = "data",
  iso3 = "KEN",
  iso3_column = "ISO3"
  )

# Load all shapefiles in a folder
shp_all <- elsar_load_data(
  file_name = NULL,
  file_path = "data/shapes"
  )

# Load PostGIS table
pg <- elsar_load_data(
  file_name = "postgres",
  file_lyr = "admin",
  iso3 = "KEN",
  iso3_column = "iso3",
  pg_connection = list(
      host = "localhost",
      dbname = "gis",
      port = 5432,
      user = "me",
      password = "pass"
      )
   )
}
}
