% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsar_plot_distribution.R
\name{elsar_plot_distribution}
\alias{elsar_plot_distribution}
\title{Function to create basic histogram with categorical data}
\usage{
elsar_plot_distribution(
  type = "ggplot",
  categorical = TRUE,
  raster_df = NULL,
  col_interest = NULL,
  bin_number = 10,
  custom_palette = NULL,
  show_legend = FALSE,
  color_map = "viridis"
)
}
\arguments{
\item{type}{A character denoting whether "ggplot" or "tmap" is being used.}

\item{categorical}{logical. Whether data is categorical (TRUE) or continuous (FALSE).}

\item{raster_df}{A \code{df} object of the data to include in the histogram.}

\item{col_interest}{A character of the column of interest. Uses "interval" if nothing is provided.}

\item{bin_number}{An integer value of how many bins to split the data into (default = 10).}

\item{custom_palette}{An optional custom palette for plotting. Default uses the \code{viridis} package.}

\item{show_legend}{logical. Whether to include legend in plot or not. Default is FALSE.}

\item{color_map}{The name of the \code{viridis} palette to be used. Default is "viridis".}
}
\value{
A \code{list} of a \code{ggplot} or \code{tmap} object and a \code{SpatRaster} with the new background data.
}
\description{
Function to create basic histogram with categorical data
}
