% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_rasterise.R
\name{exact_rasterise}
\alias{exact_rasterise}
\title{Efficient Rasterization of polygon features with Coverage Fraction, based on some attribute}
\usage{
exact_rasterise(features, attribute, pu_layer, fun = sum, cores = 4)
}
\arguments{
\item{features}{A spatial feature object (e.g., \code{sf} or \code{SpatVector}) containing
the vector features to be rasterised.}

\item{attribute}{A string representing the column name in \code{features} used to assign
values to each raster cell.}

\item{pu_layer}{A raster layer (planning unit) that defines the resolution and extent
for rasterization.}

\item{fun}{A function used to aggregate values in overlapping layers (default is \code{sum}).
Other functions, like \code{mean} or \code{max}, can also be used.}
}
\value{
A raster stack with each layer representing the rasterized attribute for each
feature, based on some attribute. If multiple features are present, the specified aggregation function
is applied across the stack to produce a single layer.
}
\description{
Rasterises vector features using the attribute values and their actual coverage
fraction over a planning unit (PU) layer. This method is optimized for speed and
precision, allowing for faster rasterisation than standard \code{rasterize()} functions,
where polygons in a dataset may overlap or have attributes to map.
}
\examples{
\dontrun{
# Rasterize using the 'value' attribute with sum aggregation
exact_rasterise(features = my_features, attribute = "value", pu_layer = my_raster, fun = sum)
}
}
