% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coverage}
\alias{get_coverage}
\title{Calculate country areal coverage (as a percentage) of an input binary layer, likely representing a zone.}
\usage{
get_coverage(zone_layer, pu_layer)
}
\arguments{
\item{zone_layer}{A binary \code{spatRast} representing a zone or protected area type feature.}

\item{pu_layer}{A binary `spatRast`` representing the planning units.}
}
\value{
A vector of areal targets of length 1.
}
\description{
Calculate country areal coverage (as a percentage) of an input binary layer, likely representing a zone.
}
\examples{
\dontrun{
   get_coverage(PA, pu)
   get_coverage(zone_protect, pu)
   get_coverage(zone_manage, pu)
}
}
\seealso{
\code{\link[terra:global]{terra::global()}} which this function wraps.
}
