% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn_forests.R
\name{get_iucn_forests}
\alias{get_iucn_forests}
\title{Extract and Rasterise IUCN GET Forest Ecosystems}
\usage{
get_iucn_forests(
  iucn_get_directory,
  iso3,
  pus,
  iucn_get_prefixes = "T",
  include_minor_occurrence = TRUE,
  output_path = NULL
)
}
\arguments{
\item{iucn_get_directory}{Character. Path to the directory containing the IUCN \code{.gpkg} files.}

\item{iso3}{Character. ISO3 country code, used for naming and passed to \code{make_normalised_raster()}.}

\item{pus}{SpatRaster. The planning units raster (terra object) over which to compute forest coverage.}

\item{output_path}{Optional character. If provided, the output raster is saved to this directory as a GeoTIFF.}

\item{include_minor_occurence}{Logical. Whether to include polygons marked as minor occurrence (default = TRUE).}
}
\value{
A normalized \code{SpatRaster} showing fractional IUCN forest coverage across planning units.
}
\description{
This function extracts forest ecosystem vector layers from IUCN GET \code{.gpkg} files
(those starting with \code{"T"}), merges and dissolves them, and calculates the proportional
coverage of forests across each planning unit using \code{exactextractr}. The result is then
normalized to a 0–1 scale using \code{elsar::make_normalised_raster()}.
}
\details{
Optionally, the result can be written to a Cloud-Optimized GeoTIFF file.
}
\examples{
\dontrun{
pus <- terra::rast("data/pus.tif")
forest_layer <- get_iucn_forests(
  iucn_get_directory = "data/iucn_layers",
  pus = pus,
  iso3 = "KEN",
  output_path = "outputs"
)
}
}
