% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{initialize_earthengine}
\alias{initialize_earthengine}
\title{Authenticate and initialize Earth Engine}
\usage{
initialize_earthengine(gee_project)
}
\arguments{
\item{gee_project}{GEE cloud project ID}
}
\value{
A list containing the \code{ee} module and temporary environment name (if created)
}
\description{
Sets up a Python environment and initializes the Earth Engine API
}
\keyword{internal}
