% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boundary.R
\name{make_boundary}
\alias{make_boundary}
\title{Create a boundary of the planning region}
\usage{
make_boundary(
  boundary_in,
  input_type = "sf",
  limit_to_mainland = FALSE,
  col_name = NULL,
  filter_out = 0,
  do_project = FALSE,
  iso3_column = "iso_sov1",
  iso3,
  output_path
)
}
\arguments{
\item{boundary_in}{A file containing the boundary information. Can be \code{sf} or \code{SpatRaster}}

\item{input_type}{A string that is either "sf" or "SpatRaster" (default is "sf").}

\item{limit_to_mainland}{Logical. Limits the extent of the data to mainland.}

\item{col_name}{A string of the column containing the actual extent of the planning region (not outside area). Can be \code{NULL}.}

\item{filter_out}{A value representing the outside area in the data (e.g. \code{0})}

\item{do_project}{Logical. \code{TRUE}if custom projection for planning region is wanted.}

\item{iso3_column}{A string of the name of where iso3 information can be found in a dataset. Only needed when do_project = TRUE.}

\item{iso3}{The iso3 country code (character) of the country of interest. Only needed when do_project = TRUE.}

\item{output_path}{An optional output path for the created file. Only needed when do_project = TRUE.}
}
\value{
\code{sf} object of the boundary of the planning region
}
\description{
Create a boundary of the planning region
}
\examples{
\dontrun{
boundary <- make_boundary(
  boundary_src = "pu_nepal_450m.tif",
  data_path = path_to_data,
  input_type = "raster", col_name = "pu_nepal_450m"
)
}
}
