% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_custom_projection.R
\name{make_custom_projection}
\alias{make_custom_projection}
\title{Create a custom projection based on the planning region}
\usage{
make_custom_projection(boundary, output_path, iso3_column = "iso_sov1", iso3)
}
\arguments{
\item{output_path}{An optional output path for the created file.}

\item{iso3_column}{Only relevant when \code{ìnput_type} "postgres" is selected. A string of the name of where iso3 information can be found in a dataset.}

\item{iso3}{The iso3 country code (character) of the country of interest.}

\item{boundary`sf`}{object of the boundary of the planning region. Should match iso3 country code.}
}
\value{
A \code{wkt} file centred on the planning region
}
\description{
Create a custom projection based on the planning region
}
\examples{
\dontrun{
boundary <- make_boundary(boundary_src = "pu_nepal_450m.tif",
data_path = path_to_data,
input_type = "raster", col_name = "pu_nepal_450m")
wkt <- make_custom_projection(boundary = boundary, output_path = outputPath, iso3 = "NPL")
boundary_proj <- sf::st_transform(boundary, crs = sf::st_crs(wkt))
}
}
