% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manage_zone.R
\name{make_manage_zone}
\alias{make_manage_zone}
\title{Generate Sustainable Management Zone}
\usage{
make_manage_zone(
  iso3,
  pus,
  mf_raster = NULL,
  agri_raster = NULL,
  built_raster = NULL,
  hfp_raster = NULL,
  hfp_threshold = 10,
  agriculture_lulc_value = 4,
  built_area_lulc_value = 7,
  filter_patch_size = TRUE,
  min_patch_size = 10,
  output_path = NULL,
  threads = TRUE
)
}
\arguments{
\item{iso3}{A character string representing the ISO3 country code (e.g., "CHL" for Chile).}

\item{pus}{A \code{SpatRaster} defining the planning units grid.}

\item{mf_raster}{A \code{SpatRaster} of Managed Forests (default: \code{NULL}).}

\item{agri_raster}{A \code{SpatRaster} of land use/land cover (LULC) data (default: \code{NULL}) filtered for agricultural areas. Can be processed or unprocessed (if unprocessed: can be the same input as built_raster)}

\item{built_raster}{A \code{SpatRaster} of land use/land cover (LULC) data (default: \code{NULL}) filtered for built-up areas. Can be processed or unprocessed (if unprocessed: can be the same input as agri_raster)}

\item{hfp_raster}{A \code{SpatRaster} of the Human Footprint Index (HFP) (default: \code{NULL}).}

\item{hfp_threshold}{A numeric threshold for identifying human impact areas in HFP (default: \code{10}).}

\item{agriculture_lulc_value}{The LULC classification value representing agricultural areas, based on the ESRI 10m LULC dataset (default: \code{4}).}

\item{built_area_lulc_value}{The LULC classification value representing built-up areas, based on the ESRI 10m LULC dataset (default: \code{7}).}

\item{filter_patch_size}{Logical; if \code{TRUE}, small patches below \code{min_patch_size} are removed (default: \code{TRUE}).}

\item{min_patch_size}{The minimum patch size (in pixels) to retain during filtering (default: \code{10}).}

\item{output_path}{A character string specifying the directory to save output rasters (default: \code{NULL}, i.e., not saved).}

\item{threads}{Logical; whether to enable multi-threaded processing where available (default: \code{TRUE}).}

\item{lulc_threshold}{A numeric threshold for agriculture and built-up areas classification (default: \code{0.1}).}
}
\value{
A \code{SpatRaster} object representing degraded areas for restoration, classified as \code{1} (degraded) and \code{0} (not degraded).
If \code{output_path} is provided, the raster is saved as a Cloud-Optimized GeoTIFF (COG).
}
\description{
Manage zone: 0<HFP<10 (middle 60\% of whole country), include all managed forests and agricultural areas, NOT include urban areas#####
}
\examples{
\dontrun{
restore_zone <- make_restore_zone(
  iso3 = "CHL",
  pus = planning_units,
  sdg_degradation_input = sdg_raster,
  lulc_raster = landcover_raster,
  hii_input = hii_raster,
  output_path = "path/to/output"
)
}
}
