% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_normalised_raster.R
\name{make_normalised_raster}
\alias{make_normalised_raster}
\title{Function to extract general raster data}
\usage{
make_normalised_raster(
  raster_in,
  pus,
  iso3,
  invert = FALSE,
  name_out,
  output_path = NULL
)
}
\arguments{
\item{raster_in}{A raster file that contains the data to be put into right format}

\item{pus}{A raster file that contains the reference spatial extent, crs etc.in form of the planning units}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{invert}{Logical. If TRUE, highest values in the original dataset should be valued the lowest in the prioritisation.}

\item{name_out}{A string with the data name that will be used for the output \code{tif}file}

\item{output_path}{An optional output path for the created file.}
}
\value{
A \code{raster}file that has been aligned and normalised
}
\description{
\code{make_normalised_raster()}allows you to align, normalise and save a raster file. This is applicable for the following data
This is applicable for the following data sets: certified forests, drought risk, flood risk, intact wilderness area, soc difference, threatened species richness, voc, wad convergence evidence
}
\examples{
boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd",
  do_project = TRUE
)

pus <- make_planning_units(boundary_proj = boundary_proj,
                           pu_size = NULL,
                           pu_threshold = 8.5e5,
                           limit_to_mainland = FALSE)
wad_dat <- get_wad_data()

wadOut <- make_normalised_raster(raster_in = wad_dat,
                                 pus = pus,
                                 iso3 = "NPL")
}
