% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_normalised_raster.R
\name{make_normalised_raster}
\alias{make_normalised_raster}
\title{Function to extract general raster data}
\usage{
make_normalised_raster(
  raster_in,
  pus,
  iso3,
  invert = FALSE,
  name_out,
  write_file = TRUE,
  output_path
)
}
\arguments{
\item{raster_in}{A raster file that contains the data to be put into right format}

\item{pus}{A raster file that contains the reference spatial extent, crs etc.in form of the planning units}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{invert}{Logical. If TRUE, highest values in the original dataset should be valued the lowest in the prioritisation.}

\item{name_out}{A string with the data name that will be used for the output \code{tif}file}

\item{write_file}{A logical command on whether the file will be saved rather than given as an output variable (Default: TRUE)}

\item{output_path}{An output path for the created file.}
}
\value{
A \code{raster}file that has been aligned and normalised
}
\description{
\code{make_normalised_raster()}allows you to align, normalise and save a raster file. This is applicable for the following data
This is applicable for the following data sets: certified forests, drought risk, flood risk, intact wilderness area, soc difference, threatened species richness, voc, wad convergence evidence
}
