% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_protected_areas.R
\name{make_protected_areas}
\alias{make_protected_areas}
\title{Function to create existing protected area raster file.}
\usage{
make_protected_areas(
  from_wdpa = TRUE,
  iso3,
  download_path = NULL,
  sf_in,
  status = c("Established", "Inscribed", "Designated"),
  pa_def = 1,
  designation_mab = FALSE,
  buffer_points = TRUE,
  area_column = "REP_AREA",
  pus,
  output_path = NULL
)
}
\arguments{
\item{from_wdpa}{logical. If \code{TRUE}, downloads data from the \pkg{wdpar} package. For more information on the data or package, please consult the \href{https://prioritizr.github.io/wdpar/}{online package documentation}.}

\item{iso3}{A string of the iso3 name of the data (country name)}

\item{download_path}{A path that the wdpa data is saved to.}

\item{sf_in}{An \code{sf} object and alternative to downloading the data from the \pkg{wdpar} package. Only needed when from_wdpa is \code{FALSE}.}

\item{status}{A vector containing which status of protected area should be included. Based on the STATUS field of the wdpa database. Default is \code{c("Established", "Inscribed", "Designated")}.}

\item{pa_def}{A value or list of values containing which pa definition (1 = protected area; 0 = OECM (not supported yet)) to include. Default is \code{1}.}

\item{designation_mab}{logical. If \code{FALSE}, excludes UNESCO MAB areas.}

\item{buffer_points}{logical. Only relevant when \code{"POINT"} or \code{"MULTIPOINT"} geometries exist in the data. If \code{TRUE}, creates a circular buffer around \code{"POINT"} data based on area information data that is then used as polygon data needed for \pkg{exactextractr} calculations.}

\item{area_column}{A string of the column name with the area information needed for buffer calculations.}

\item{pus}{A raster file that contains the reference spatial extent, crs etc.in form of the planning units.}

\item{output_path}{An optional output path for the created file.}
}
\value{
A \code{raster}file that has been aligned to the planning units
}
\description{
Function to create existing protected area raster file.
}
\examples{
\dontrun{
boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd",
  do_project = TRUE
)

pus <- make_planning_units(
  boundary_proj = boundary_proj,
  pu_size = NULL,
  pu_threshold = 8.5e5,
  limit_to_mainland = FALSE
)

current_pas <- make_protected_areas(
  iso3 = "NPL",
  download_path = here::here(),
  buffer_points = FALSE,
  pus = pus
)
}
}
