% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_urban_greening_opportunities.R
\name{make_urban_greening_opportunities}
\alias{make_urban_greening_opportunities}
\title{Create Urban Greening Opportunities Raster}
\usage{
make_urban_greening_opportunities(
  ndvi_raster,
  lulc_raster,
  sdei_statistics,
  pus,
  iso3,
  name_out,
  output_path = NULL,
  cores = 4
)
}
\arguments{
\item{ndvi_raster}{A \code{SpatRaster} representing the NDVI data.}

\item{lulc_raster}{A \code{SpatRaster} of the land use/land cover (LULC) classification.}

\item{sdei_statistics}{A \code{sf} object represetning GHS boundaries and SDEI extreme
urban heat data.}

\item{pus}{A \code{SpatRaster} defining the planning unit (PU) grid.}

\item{iso3}{A character string representing the ISO3 country code.}

\item{output_path}{A character string specifying the output directory (optional).}

\item{cores}{A number allocating the availabe cores for \code{terra} tools that allow
multi-core processing (if available on your machine). Defaults to 4.}
}
\value{
A \code{SpatRaster} representing urban greening opportunities.
}
\description{
This function generates an urban greening opportunities raster by integrating NDVI,
land-use classifications (LULC), and extreme heat exposure from GHS data.
The process involves normalizing NDVI, extracting urban areas from LULC,
rasterizing extreme heat exposure, and combining these layers to identify potential
areas for greening interventions.
}
\examples{
\dontrun{
ndvi_raster <- elsar::elsar_load_data(file_path = ".", file_name = "mod13q1_2022_ndvi_wgs84.tif", file_type = "tiff")

lulc_raster <- elsar::elsar_load_data(file_path = ".", file_name = "esri_10m_lulc_2023_NPL.tif", file_type = "tiff")

sdei_statistics <- sf::read_sf(
  "~/sdei-high-res-daily-uhe-1983-2016-shp/wbgtmax30_join.gpkg") |>
  sf::st_set_geometry(NULL) |>
  dplyr::group_by(ID_HDC_G0) |>
  dplyr::summarise(avg_intens = mean(avg_intens)
  )

urban_green_opportunities <- make_urban_greening_opportunities(
  ndvi_raster = ndvi,
  lulc_raster = lulc,
  sdei_statistics = sdei_statistics,
  pus = planning_units,
  iso3 = "NPL",
  )

terra::plot(urban_green_opportunities)
}
}
