% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wetlands_ramsar.R
\name{make_wetlands_ramsar}
\alias{make_wetlands_ramsar}
\title{Function to make standardised Wetlands and Ramsar data}
\usage{
make_wetlands_ramsar(
  ramsar_in = NULL,
  wetlands_in = NULL,
  pus,
  iso3_in,
  buffer_points = TRUE,
  area_column = "Area (ha)",
  nQuadSegs = 50,
  return_all = FALSE
)
}
\arguments{
\item{ramsar_in}{An \code{sf} file that contains the ramsar data to be put into right format}

\item{wetlands_in}{A \code{SpatRaster} file with the wetlands information}

\item{pus}{A \code{SpatRaster} file that contains the reference spatial extent, crs etc.in form of the planning units}

\item{iso3_in}{A string of the iso3 name of the data (country name)}

\item{buffer_points}{logical. Only relevant when \code{"POINT"} or \code{"MULTIPOINT"} geometries exist in the data. If \code{TRUE}, creates a circular buffer around \code{"POINT"} data based on area information data that is then used as polygon data needed for \pkg{exactextractr} calculations.}

\item{area_column}{A string of the column name with the area information needed for buffer calculations.}

\item{nQuadSegs}{An integer specifying the number of segments to use for buffering. Default is 50.}

\item{return_all}{Logical. Whether to return a \code{SpatRaster} with the combined wetlands and raster layer, as well as the individual layers}
}
\value{
A \code{SpatRaster} file that has been aligned and normalised
}
\description{
Function works with either Ramsar or Wetlands data or both.
}
