% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_helpers.R
\name{merge_tiles}
\alias{merge_tiles}
\title{Merge downloaded tiles into a single COG}
\usage{
merge_tiles(local_path, output_file, datatype)
}
\arguments{
\item{output_file}{Final merged COG file path}

\item{datatype}{Output raster datatype}

\item{tiles}{List of GeoTIFF file paths}
}
\value{
A \code{SpatRaster} object
}
\description{
Merge downloaded tiles into a single COG
}
\keyword{internal}
